microservice_project_template/src/microservice_template/observability/tracing.py
"""OpenTelemetry tracing configuration."""

from __future__ import annotations

from microservice_template.config import AppSettings
from opentelemetry import trace
from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
from opentelemetry.sdk.resources import Resource
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor


def configure_tracer(settings: AppSettings) -> TracerProvider | None:
    """Configure and register the global tracer provider if tracing is enabled."""

    if not settings.tracing_enabled:
        return None

    resource = Resource.create(
        {
            "service.name": settings.service_name,
            "service.version": settings.version,
            "deployment.environment": settings.environment,
        }
    )

    provider = TracerProvider(resource=resource)
    exporter = OTLPSpanExporter(endpoint=settings.tracing_endpoint, insecure=True)
    provider.add_span_processor(BatchSpanProcessor(exporter))
    trace.set_tracer_provider(provider)
    return provider
